マジックナンバー一部消去 この後敵の再生成を修正、実装する

+++ MyGAME/GameTemplate/Game/EnemyAtack.h
@@ -5,7 +5,7 @@ class EnemyAtack
 private:
 	Vector3 m_enemyPos = Vector3::Zero;
 	Vector3 m_playerPos = Vector3::Zero;
-	int counter = 0;
+	float counter = 0.0f;
 	bool startAtack = false;
 public:
 	void SetEnemyPos(Vector3 enemyPos) { m_enemyPos = enemyPos; }
